# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Документация API вашего приложения",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Основной API сервер"
        }
    ],
    "paths": {
...
        "/api/organizations": {
            "get": {
                "tags": [
                    "Организации"
                ],
                "summary": "Получить список всех организаций",
                "operationId": "42efb13746ba71d73c330585f09ecef8",
                "responses": {
                    "200": {
                        "description": "Список организаций",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Organization"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Организации"
                ],
                "summary": "Создать новую организацию",
                "operationId": "2d849ce559213546c5ec62ee8ea255db",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "building_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "ООО Рога и Копыта"
                                    },
                                    "building_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "phones": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "+7 999 123-45-67"
                                        }
                                    },
                                    "activities": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 5
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Организация успешно создана"
                    },
                    "422": {
                        "description": "Ошибка валидации данных"
                    }
                }
            }
        },
        "/api/organizations/{id}": {
            "get": {
                "tags": [
                    "Организации"
                ],
                "summary": "Получить организацию по ID",
                "operationId": "68e75650b204083886e125e02aaa2f4b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID организации",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация об организации"
                    },
                    "404": {
                        "description": "Организация не найдена"
                    }
                }
            },
            "put": {
                "tags": [
                    "Организации"
                ],
                "summary": "Обновить организацию по ID",
                "operationId": "a6863f1e78baca329a5507e9d3d8e32d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID организации",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "ООО Обновлённое Имя"
                                    },
                                    "building_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "phones": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "activities": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Организация обновлена"
                    },
                    "404": {
                        "description": "Организация не найдена"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Организации"
                ],
                "summary": "Удалить организацию по ID",
                "operationId": "437c30a09c2dda264e402fb5ca96e873",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID организации",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Организация удалена"
                    },
                    "404": {
                        "description": "Организация не найдена"
                    }
                }
            }
        },
        "/api/organizations/nearby": {
            "get": {
                "tags": [
                    "Организации"
                ],
                "summary": "Найти организации поблизости",
                "operationId": "2c4aaebf1645ecdcad5af01f7b635407",
                "parameters": [
                    {
                        "name": "lat",
                        "in": "query",
                        "description": "Широта",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "example": 55.751244
                        }
                    },
                    {
                        "name": "lng",
                        "in": "query",
                        "description": "Долгота",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "example": 37.618423
                        }
                    },
                    {
                        "name": "radius",
                        "in": "query",
                        "description": "Радиус поиска (в км, по умолчанию 5)",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Организации поблизости",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Organization"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "radius_km": {
                                                    "type": "number"
                                                },
                                                "found_buildings": {
                                                    "type": "integer"
                                                },
                                                "found_organizations": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/organizations/search": {
            "get": {
                "tags": [
                    "Организации"
                ],
                "summary": "Поиск организаций по названию",
                "operationId": "40b8b5635dc20a80567117691e82d2e0",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Поисковая строка (часть названия организации)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Рога"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Результаты поиска",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Organization"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Не указана поисковая строка"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Activity": {
                "title": "Activity",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 5
                    },
                    "name": {
                        "type": "string",
                        "example": "Еда"
                    },
                    "parent_id": {
                        "type": "integer",
                        "example": null,
                        "nullable": true
                    },
                    "_lft": {
                        "type": "integer",
                        "example": 1
                    },
                    "_rgt": {
                        "type": "integer",
                        "example": 8
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Building": {
                "title": "Здание",
                "description": "Здание, где расположена организация",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 10
                    },
                    "address": {
                        "type": "string",
                        "example": "г. Москва, ул. Ленина, 10"
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float",
                        "example": 55.7558
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float",
                        "example": 37.6173
                    }
                },
                "type": "object"
            },
            "Organization": {
                "title": "Организация",
                "description": "Модель организации",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "ООО Ромашка"
                    },
                    "building_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "phones": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "+7 (900) 123-45-67"
                        }
                    },
                    "activities": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Activity"
                        }
                    },
                    "building": {
                        "$ref": "#/components/schemas/Building"
                    }
                },
                "type": "object"
            }
        }
    },
...
}